Change the default Timezone handling to UTC
In the Drupal Community Documentation it is
mentioned: Time zone handling -  UTC: When
entering data into the field, the data entered is
assumed to be in UTC time zone. When the data is
saved to the database, it is converted to UTC
(e.g. no conversion necessary). When retrieved
from the database, the data is converted to the
Site's time zone for anonymous users or the User's
time zone for logged in users when
User-configurable time zones is enabled.